Add a generic test for all services
The test runs under systemd with runit-run pkg; for each service
defined in the package it search for the package that ships the
corresponding bin and install it; then it disable systemd related
units and enable the service under runit. After that, the following
tests are performed:

* check that the service is up, trying hard to detect races where the
  service is temporary up (for 1s or less);
* check that the appendant log (if any) is not empty
* if the service fails and has a log, the last 20 lines are printed

Given the large number of servies tested and the fact that testing
under systemd is not optimal the test is marked as flaky.
(on testing under systemd: as serviecs are first started under
 systemd, directories and users may be created and capabilities
 may be set and this may cause services to fail on runit under
 systemd but not with runit-ini or vice-versa)