dhclient: ship as disabled by default
it appears that dhclient is started automatically by some network
software; other inits also do not appear to have a services defined
for dhclient. So shipping as disabled by default is the least
surprise approach and also avoid duplicate instances.
This change is applied to new install of this package, on
upgrade the old setting is retained.