Provides mini-httpd-run
Provides mini-httpd-run, since this package now includes a runscript
for the service equivalent to the one in mini-httpd-run package.
There is no need for a Conflict for now since this package installs
under /usr while the other pacakge installs in /etc/sv/ so the two
packages can be installed at the same time.