override rsyslog logrotate file
ship a copy of /etc/logrotate.d/rsyslog namend so that it's processed
early than the one provided by rsyslog package; according to
logrotate(8), section frequency/weekly, rotation happens only if files
were not already rotated in the current week, so if the copy of rsyslog
file is processed early it overrides the one from rsyslog package.
The copy of rsyslog logrotate file is handled with a noawait trigger.
The copy is removed on purge.
Attempts to fix this in a proper and cooperative way failed,
see #1031399 and #1079270

Closes: #1079268