exim4: fix exim4 launch with trixie default env
Since the change to the default environment shipped after bookworm in
<https://salsa.debian.org/exim-team/exim4/-/commit/ea38b9fddb71d1b5e9a125ca3a2ce7b7cfae0898>
the runscript for exim4 will by default attempt to start the daemon with
invalid command line arguments (namely missing queue runner interval):

2024-11-09_14:17:13.18129 /usr/sbin/exim4 -bd -d -q
2024-11-09_14:17:13.18358 exim: incompatible command-line options or arguments

This patch fixes this by using the new EXIMSERVICE default and providing
a default for that variable matching other init systems.

Additionally, this runscript omits an invocation of update-exim4.conf to
assemble the debian autogenerated config where applicable. This patch
does so on launch and on HUP.

Defaults from pre-trixie configurations are used by preference if present.

Closes: #1087204