improve binpath test
try to error only when the path is wrong, warning when binpath is
not found.
This prevent failures for binaries that are arch specific (acpi_fakekey,
zcfan..) but should still catch cases like #1069075, where the bin
is found but the path is different from the one that is in the control
file.
Also, skip empty lines, which seems to create issues (test timeout)
in arm64 arch.