fix elogind for bookworm-trixie upgrade
force upgrade the path of bin metafiles in /etc/sv/elogind/.meta,
if any; backup /etc/sv/elogind/run if any, then replace it with
the pacakge version. this needs a NEWS file entry.

elogind changed path; path for the elogind runscript in this
package was upgraded as well, but local version in /etc/sv/ still
has the wrong one.
For contest, this package in bookworm used to copy services into
/etc/sv/ but now we no longer do this and use the one under /usr/
exactly for the reason that scripts under /etc/sv were not upgraded
automatically even if not modified.
The issue with elogind is bad since affects both the boot process
and desktop stuff (see also #1069075)

Closes: #1095025